bot_6/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
asyncpg==0.24.0
aioredis==2.0.0
python-binance==1.0.15
okx-python-sdk==0.4.0
pyyaml==5.4.1
numpy==1.21.2
pandas==1.3.3